do not catch exceptions in action coderefs
Previous behaviour was to catch exceptions in action coderefs, leaving
the user with no clue if the action did something wrong, e.g., using a
nonexistent method on the iterator object. Now the user is required to
wrap potentially dangerous code in an 'eval' block (the test suite
contains some examples). Although this may seem like a burden on the
user, the use of extra 'eval' blocks does not seem to be required often
(judging from the test suite, at least).