test bad value support early in the test suite
As explained in the previous commit, PDL::NDBin needs bad value support
to work. This is tested by running a test early in the test suite.
There's probably a better way to do this (like a Dist::Zilla plugin?),
but until I figure out how to do it, this will have to do.