require axis specification to be key=>value pairs
Axis specifications passed to the constructor new() of PDL::NDBin used
to be a bare set of three numbers, specifying step size, minimum value,
and number of bins. In preparation of the transfer of the sophisticated
argument processing from ndbin() to new(), we now require the three
numbers to be labelled. Additionally, I believe this change also
improves the maintainability of the calling code, as the axis
specifications are now labelled explicitly (i.e., you don't have to
remember what the numbers stand for).