simplify and optimize PDL::NDBin::Iterator
Profiling indicated that a very significant amount of time is spent in
Iterator.pm, and particularly in next1() and next(). This commit both
simplifies and optimizes the iteration: there is now a single method
advance() which has been optimized somewhat. Some preliminary testing
indicated a speed-up with a factor of three. The performance bottleneck
in advance() may very well be the reason that PDL::NDBin performs so
poorly on very large bin numbers. (Note that a lot of time is spent on
looking for the next active variable, so advance() is bound to waste a
lot of time.)