do not compare automatic number of bins with hist()
Disabled two tests that tested whether the number of bins determined by
PDL::NDBin, when it was not supplied, matches the number of bins
determined by hist(). The tests were giving test failures with the
newest development releases of PDL (2.4.11_007 and 2.4.11_008 among
them), because the bin number determination has been changed. Since
PDL's behaviour was not documented, relying on it was not a good idea.
Hence, the tests were disabled. Updated the user documentation to match
the code.