Use PDL::Core instead of PDL::Lite in t/07-grid.t
(Some) older versions of PDL did have a bug in PDL::Lite, where function
'pdl' was not exported as it should. This commit tries to work around
that (see https://github.com/PDLPorters/pdl/issues/224). Recent versions
of PDL don't have this problem, but I'm still seeing a lot of test
failure reports, so I'm issuing this fix in the hope of improving the
CPAN Test matrix stats :)